Refactor configuration handling by replacing module with externaldownload configuration file and updating import paths in copy scripts
