+++ scripts/copy-original-images.cjs
@@ -1,4 +1,4 @@
-const config = require('../src/config/module.cjs');
+const config = require('../src/config/externaldownload.cjs');
 if (!config.enableCopyOriginalImages) {
   console.log('Originalbild-Download ist deaktiviert. Überspringe Skript.');
   process.exit(0);
